Introduce RequestController
RequestController extracts and abstracts some common functionality
of executing a request
