+++ http-client/src/main/java/io/airlift/http/client/jetty/RequestSizeListener.java
@@ -1,0 +1,20 @@
+package io.airlift.http.client.jetty;
+
+import java.nio.ByteBuffer;
+
+class RequestSizeListener
+        implements org.eclipse.jetty.client.api.Request.ContentListener
+{
+    private long bytes;
+
+    @Override
+    public void onContent(org.eclipse.jetty.client.api.Request request, ByteBuffer content)
+    {
+        bytes += content.remaining();
+    }
+
+    public long getBytes()
+    {
+        return bytes;
+    }
+}
